requirements.txt
kfp==2.1.0
google-cloud-aiplatform==2.18.0
scikit-learn
numpy
joblib
pytest
